Refactor project structure; separate main.rs in main.rs and app.rs
Move all emulator logic to chip8 subdirectory. Divide functionality
from main.rs. The responsibilty of main.rs is now to parse CLI args
and create a config struct instance as well as run the Chip8App.
The responsibilty of app is to create the window, interact with
the emulator core and execute the main loop.